+++ CHANGELOG.md
@@ -18,6 +18,8 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 ### Changed
 
 - Modified create booking code to allow conflicts if resource.acceptConflict is true.
+- Changed how location email and name is found in user bookings.
+- Set displayName from resource database when set.
 
 ## [1.0.4] - 2023-05-03